requirements.txt
fastapi
uvicorn
tensorflow
numpy
pillow
scikit-learn
joblib
python-multipart